Feature/#8 get certain news

+++ src/main/java/com/dunice/Vitushkin_Advanced_REST_Server/response/PageableResponse.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 public class PageableResponse<T> {
     private T content;
-    private Integer numberOfElements;
+    private Long numberOfElements;
 }
